Avec de l'espoir, fin de la mise en place du squelette du design. A besoin de nombreux fixes.

+++ sources/Serveur.java
@@ -2,7 +2,7 @@ package trollhammer;
 
 public class Serveur {
     
-    private int date;
+    private long date;
     private String dernier_enchérisseur;
     private Onglet mode;
     private int nouveau_prix;
